[lua] Add interaction support and helper functions
Account for tick time in vanaTime enum

Fix obo for NextJstWeek

Add reference for vanaTime Values

Fix NextGameTime logic

Fix trailing whitespace

+++ src/map/lua/luautils.h
@@ -163,7 +163,7 @@ namespace luautils
     uint32 GetSystemTime();
     uint32 JstMidnight();
     uint32 JstWeekday();
-    uint32 NextGameTime(uint16 intervalSeconds);
+    uint32 NextGameTime(uint32 intervalSeconds);
     uint32 NextJstWeek();
     uint32 VanadielTime();          // Gets the current Vanadiel Time in timestamp format (SE epoch in earth seconds)
     uint8  VanadielTOTD();          // текущее игровое время суток
